# internal/api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: booth-spark job-submission API
  version: "1.0.0"
  summary: Submit and manage Spark applications on Project Booth.
  description: |
    The published interface of booth-spark (docs/design-v0.md item 6). Reached through booth-core's
    gateway at `/modules/spark/v1/...` with a bearer token and the `X-Workspace` header: a person's
    OIDC token, or booth-core's workload token for an unattended caller such as a pipeline task
    (ADR 0059).

    Versioning: within v1, changes are additive only (new optional fields, new endpoints).
    Anything breaking is `/v2`, served alongside `/v1` for at least one minor release.

    Who may do what (docs/design-v0.md item 3): anyone in the workspace sees its runs; submitting
    needs `submit.minRole` (editor by default, owner allowed; never viewer); stopping a run or
    reading its logs needs its submitter, a workspace owner or a platform operator. A run's Spark UI
    opens in the shell for its submitter only (`uiPath`).

    Sessions are their submitter's and their workspace owners': nobody else sees them, and only the
    submitter runs statements in one. A session with nothing waiting or running and no new
    statement for its idle timeout is stopped, as is every session after its maximum lifetime.

    Not in this version: data access (`dataAccess`,
    `main.python`, `main.jar` are refused with 422), and any data access for runs submitted with a
    workload token (ADR 0110 ruling 5).
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /modules/spark/v1
    description: Through booth-core's gateway
security:
  - bearer: []
tags:
  - name: identity
    description: Who the caller is, as this module derives it
  - name: applications
    description: Batch Spark applications, one namespace per run
  - name: sessions
    description: Long-lived Spark sessions that take statements, one namespace each; stopped when idle
paths:
  /me:
    get:
      tags: [identity]
      operationId: getMe
      summary: The caller, as booth-spark derived it from the verified token
      parameters:
        - $ref: "#/components/parameters/Workspace"
      responses:
        "200":
          description: The caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "503": {$ref: "#/components/responses/Error"}
  /applications:
    get:
      tags: [applications]
      operationId: listApplications
      summary: The workspace's runs, newest first (at most 200)
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - name: state
          in: query
          required: false
          description: Only runs in this state
          schema: {$ref: "#/components/schemas/State"}
        - name: mine
          in: query
          required: false
          description: Only the caller's own runs
          schema: {type: boolean}
      responses:
        "200":
          description: The runs
          content:
            application/json:
              schema:
                type: object
                required: [applications]
                properties:
                  applications:
                    type: array
                    items: {$ref: "#/components/schemas/Run"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
    post:
      tags: [applications]
      operationId: submitApplication
      summary: Submit an application
      description: |
        Admission is checked first: at most `runs.maxRunning` live runs install-wide (and
        `runs.maxRunningPerWorkspace` per workspace), and every live run's largest footprint must
        fit `runs.memoryBudget`; otherwise 429.
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            A caller-chosen key (at most 200 characters). Submitting again with the same key, by the
            same caller in the same workspace, returns the first run (200) instead of starting another.
          schema: {type: string, maxLength: 200}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ApplicationSpec"}
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Run"}
        "200":
          description: Already submitted with this Idempotency-Key
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Run"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /applications/{id}:
    get:
      tags: [applications]
      operationId: getApplication
      summary: One run
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: The run
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Run"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /applications/{id}/stop:
    post:
      tags: [applications]
      operationId: stopApplication
      summary: Stop a run (its submitter, a workspace owner or a platform operator)
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
      responses:
        "202":
          description: Stop requested; the run becomes `stopped` and its namespace is deleted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Run"}
        "200":
          description: The run had already ended
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Run"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /applications/{id}/logs:
    get:
      tags: [applications]
      operationId: getApplicationLogs
      summary: The driver's log (its submitter, a workspace owner or a platform operator)
      description: Live while the run is live; afterwards, the last 1 MiB kept when it ended, until sessions.resultRetention after that (then 410 cleared).
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
        - name: tail
          in: query
          required: false
          description: Lines from the end of a live log (default 500)
          schema: {type: integer, minimum: 1, maximum: 10000}
      responses:
        "200":
          description: The log
          content:
            text/plain:
              schema: {type: string}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "410": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
  /sessions:
    get:
      tags: [sessions]
      operationId: listSessions
      summary: The caller's sessions (all of the workspace's, for a workspace owner), newest first
      parameters:
        - $ref: "#/components/parameters/Workspace"
      responses:
        "200":
          description: The sessions
          content:
            application/json:
              schema:
                type: object
                required: [sessions]
                properties:
                  sessions:
                    type: array
                    items: {$ref: "#/components/schemas/Session"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
    post:
      tags: [sessions]
      operationId: startSession
      summary: Start a session
      description: |
        Admitted like an application (runs.maxRunning, runs.memoryBudget; otherwise 429). The session
        becomes `running` once its SparkSession is ready; statements sent before then wait.
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - name: Idempotency-Key
          in: header
          required: false
          description: As for applications
          schema: {type: string, maxLength: 200}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SessionSpec"}
      responses:
        "201":
          description: Started
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "200":
          description: Already started with this Idempotency-Key
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "422": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /sessions/{id}:
    get:
      tags: [sessions]
      operationId: getSession
      summary: One session (its submitter or a workspace owner)
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: The session
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
    delete:
      tags: [sessions]
      operationId: deleteSession
      summary: Stop a session and delete its namespace (its submitter or a workspace owner)
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
      responses:
        "202":
          description: Stop requested; the session becomes `stopped` and its namespace is deleted
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "200":
          description: The session had already ended
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Session"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
  /sessions/{id}/logs:
    get:
      tags: [sessions]
      operationId: getSessionLogs
      summary: The session driver's log (its submitter or a workspace owner)
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: The log
          content:
            text/plain:
              schema: {type: string}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "410": {$ref: "#/components/responses/Error"}
        "502": {$ref: "#/components/responses/Error"}
  /sessions/{id}/statements:
    get:
      tags: [sessions]
      operationId: listStatements
      summary: The session's statements, in order (its submitter or a workspace owner)
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: The statements
          content:
            application/json:
              schema:
                type: object
                required: [statements]
                properties:
                  statements:
                    type: array
                    items: {$ref: "#/components/schemas/Statement"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
    post:
      tags: [sessions]
      operationId: runStatement
      summary: Run a statement in the session (its submitter only)
      description: |
        Statements run one at a time, in order, in the session's one SparkSession (`spark`, `sc`).
        A statement that fails ends as `error` with its traceback; the session goes on. At most 20
        statements may be waiting or running at once (429). A statement waiting or running keeps
        the session from going idle.
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/StatementRequest"}
      responses:
        "201":
          description: Queued
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Statement"}
        "400": {$ref: "#/components/responses/Error"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
        "409": {$ref: "#/components/responses/Error"}
        "429": {$ref: "#/components/responses/Error"}
  /sessions/{id}/statements/{sid}:
    get:
      tags: [sessions]
      operationId: getStatement
      summary: One statement, with its output once it has ended
      parameters:
        - $ref: "#/components/parameters/Workspace"
        - $ref: "#/components/parameters/RunID"
        - name: sid
          in: path
          required: true
          schema: {type: string}
      responses:
        "200":
          description: The statement
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Statement"}
        "401": {$ref: "#/components/responses/Error"}
        "403": {$ref: "#/components/responses/Error"}
        "404": {$ref: "#/components/responses/Error"}
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Workspace:
      name: X-Workspace
      in: header
      required: true
      description: The active workspace's slug (booth-core forwards it as X-Booth-Workspace)
      schema: {type: string}
    RunID:
      name: id
      in: path
      required: true
      schema: {type: string, pattern: "^r[a-z0-9]{11}$"}
  responses:
    Error:
      description: An error
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Error"}
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: string, description: "What went wrong, in words"}
        code:
          type: string
          description: A stable machine-readable code
          enum: [invalid, forbidden, not_found, unavailable, at_capacity, session_over, cleared, internal]
    Me:
      type: object
      required: [subject, workspace, role, operator, workload, canSubmit, submitMinRole]
      properties:
        subject: {type: string}
        displayName: {type: string}
        workspace: {type: string}
        role: {type: string, enum: [owner, editor, viewer]}
        operator: {type: boolean}
        workload: {type: boolean}
        canSubmit: {type: boolean}
        submitMinRole: {type: string, enum: [editor, owner]}
    State:
      type: string
      enum: [pending, running, succeeded, failed, stopped]
    ApplicationSpec:
      type: object
      additionalProperties: false
      required: [name, main]
      properties:
        name:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9 ._-]{0,99}$"
        main:
          type: object
          additionalProperties: false
          properties:
            inlinePython:
              type: string
              maxLength: 262144
              description: The application's Python source (PySpark), at most 256 KiB
            python:
              description: Not in this version (needs data access); refused with 422
            jar:
              description: Not in this version (needs data access); refused with 422
        args:
          type: array
          maxItems: 64
          items: {type: string, maxLength: 4096}
        resources:
          type: object
          additionalProperties: false
          properties:
            driver:
              type: object
              additionalProperties: false
              properties:
                memory: {$ref: "#/components/schemas/Memory"}
            executors:
              type: object
              additionalProperties: false
              properties:
                min: {type: integer, minimum: 0}
                max: {type: integer, minimum: 0, description: "At most runs.maxExecutors"}
                memory: {$ref: "#/components/schemas/Memory"}
        conf:
          type: object
          description: |
            Spark settings the caller may set: spark.sql.shuffle.partitions, spark.default.parallelism,
            spark.sql.adaptive.enabled, spark.sql.session.timeZone, spark.sql.ansi.enabled,
            spark.sql.execution.arrow.pyspark.enabled. Anything else is refused (400).
          additionalProperties: {type: string, maxLength: 256}
        dataAccess:
          description: Not in this version; any non-empty value is refused with 422
        maxDuration:
          type: string
          description: A Go duration (e.g. "30m"), at most runs.maxDuration, which is the default
    Memory:
      type: string
      pattern: "^[1-9][0-9]{0,5}[mg]$"
      description: A JVM heap in Spark's notation, between 512m and runs.maxMemory
    Run:
      type: object
      required: [id, kind, workspace, submitter, workload, name, state, stopRequested, createdAt, footprintMi, resources]
      properties:
        id: {type: string}
        kind: {type: string, enum: [application]}
        workspace: {type: string}
        submitter: {type: string, description: The submitter's subject}
        submitterName: {type: string}
        workload: {type: boolean, description: Submitted with a workload token}
        name: {type: string}
        state: {$ref: "#/components/schemas/State"}
        reason: {type: string, description: "Why the run is in its state, once it has ended"}
        stopRequested: {type: boolean}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time}
        finishedAt: {type: string, format: date-time}
        contentClearedAt: {type: string, format: date-time, description: "When its content (log tail; a session's statement code, output and error) was cleared, sessions.resultRetention after it ended"}
        footprintMi: {type: integer, description: "The most memory the run's pods can use at once, MiB"}
        resources:
          type: object
          required: [driverMemory, executorMemory, minExecutors, maxExecutors, maxDuration]
          properties:
            driverMemory: {type: string}
            executorMemory: {type: string}
            minExecutors: {type: integer}
            maxExecutors: {type: integer}
            maxDuration: {type: string}
        uiPath:
          type: string
          description: While running, where the run's Spark UI opens in the shell (its submitter only)
    SessionSpec:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          pattern: "^[A-Za-z0-9][A-Za-z0-9 ._-]{0,99}$"
        resources:
          type: object
          additionalProperties: false
          properties:
            driver:
              type: object
              additionalProperties: false
              properties:
                memory: {$ref: "#/components/schemas/Memory"}
            executors:
              type: object
              additionalProperties: false
              properties:
                min: {type: integer, minimum: 0}
                max: {type: integer, minimum: 0}
                memory: {$ref: "#/components/schemas/Memory"}
        conf:
          type: object
          description: The same allowlist as an application's conf
          additionalProperties: {type: string, maxLength: 256}
        dataAccess:
          description: Not in this version; any non-empty value is refused with 422
        idleTimeout:
          type: string
          description: "A Go duration, at most sessions.idleTimeout (the default)"
        maxLifetime:
          type: string
          description: "A Go duration, at most sessions.maxLifetime (the default)"
    Session:
      type: object
      required: [id, kind, workspace, submitter, workload, name, state, stopRequested, createdAt, footprintMi, resources, idleTimeout, maxLifetime]
      properties:
        id: {type: string}
        kind: {type: string, enum: [session]}
        workspace: {type: string}
        submitter: {type: string}
        submitterName: {type: string}
        workload: {type: boolean}
        name: {type: string}
        state: {$ref: "#/components/schemas/State"}
        reason: {type: string, description: "Why the session ended: deleted on request, idle, its maximum lifetime, or a failure"}
        stopRequested: {type: boolean}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time}
        finishedAt: {type: string, format: date-time}
        contentClearedAt: {type: string, format: date-time, description: "When its content (log tail; a session's statement code, output and error) was cleared, sessions.resultRetention after it ended"}
        lastActivityAt: {type: string, format: date-time}
        footprintMi: {type: integer}
        idleTimeout: {type: string}
        maxLifetime: {type: string}
        resources:
          type: object
          required: [driverMemory, executorMemory, minExecutors, maxExecutors]
          properties:
            driverMemory: {type: string}
            executorMemory: {type: string}
            minExecutors: {type: integer}
            maxExecutors: {type: integer}
        uiPath: {type: string, description: "While running, where the session's Spark UI opens in the shell (its submitter only)"}
    StatementRequest:
      type: object
      additionalProperties: false
      required: [kind, code]
      properties:
        kind: {type: string, enum: [sql, python]}
        code: {type: string, minLength: 1, maxLength: 262144}
    Statement:
      type: object
      required: [id, sessionId, seq, kind, code, state, createdAt]
      properties:
        id: {type: string}
        sessionId: {type: string}
        seq: {type: integer}
        kind: {type: string, enum: [sql, python]}
        code: {type: string}
        state: {type: string, enum: [waiting, running, available, error, cancelled]}
        output:
          description: |
            Once available: for sql, {"type": "table", "columns": [{name, type}], "rows": [[...]],
            "truncated"} (at most 1000 rows and 1 MiB); for python, {"type": "text", "stdout",
            "stderr", "truncated"} (at most 1 MiB each).
          type: object
        error: {type: string, description: "The traceback of a statement that ended as error, or why it was cancelled"}
        createdAt: {type: string, format: date-time}
        startedAt: {type: string, format: date-time}
        finishedAt: {type: string, format: date-time}